requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# API & Networking
requests>=2.28.0

# Configuration
python-dotenv>=1.0.0

# Development (Optional)
jupyter>=1.0.0
jupyterlab>=3.0.0

# Code Quality (Optional)
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0
